[GitHub #6] SQ DOC-04: fixed check for default copyright header
When using the default copyright header getCopyright() returns a string, not a list.
During the check this leads to an iteration over all characters of the string rather than the lines.
Thus not correctly checking for the copyright header, which can result in some files passing the check if they contain all of the characters of the copyright header.